resolve feedbacks
